Mas cosillas

+++ P1/src/es/pe/mdelossgavil/Poblacion/CromosomaF1.java
@@ -3,17 +3,17 @@ package es.pe.mdelossgavil.Poblacion;
 public class CromosomaF1 extends Cromosoma{
 	
 	
-	public double fenotipo() {
-		return 0.0;
+	public float fenotipo() {
+		return 0.0f;
 	}
 
-	public double funcion(double x) {
+	public float funcion(double x) {
 		// TODO calcular el fitness de la funcion correspondiente
-		return 0.0;
+		return 0.0f;
 	}
 	
 	@Override
-	public double evaluar() {
+	public float evaluar() {
 		double x;
 		x = fenotipo();
 		return funcion(x);
